requirements.txt
numpy
six
tensorflow==1.4.1
matplotlib
Pillow
requests
paho-mqtt
